+++ src/basic/hooks/useCart.ts
@@ -1,0 +1,124 @@
+import { useCallback, useEffect, useState } from 'react';
+
+import { CartItem, Product } from '../../types';
+import { getRemainingStock, calculateCartTotal } from '../utils';
+
+interface ProductWithUI extends Product {
+  description?: string;
+  isRecommended?: boolean;
+}
+
+export function useCart(
+  products: ProductWithUI[],
+  addNotification: (message: string, type?: 'error' | 'success' | 'warning') => void
+) {
+  const [cart, setCart] = useState<CartItem[]>(() => {
+    const saved = localStorage.getItem('cart');
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return [];
+      }
+    }
+    return [];
+  });
+  const [totalItemCount, setTotalItemCount] = useState(0);
+
+  // 장바구니 아이템 개수 계산
+  useEffect(() => {
+    const count = cart.reduce((sum, item) => sum + item.quantity, 0);
+    setTotalItemCount(count);
+  }, [cart]);
+
+  // localStorage에 저장
+  useEffect(() => {
+    if (cart.length > 0) {
+      localStorage.setItem('cart', JSON.stringify(cart));
+    } else {
+      localStorage.removeItem('cart');
+    }
+  }, [cart]);
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getRemainingStock(product, cart);
+      if (remainingStock <= 0) {
+        addNotification('재고가 부족합니다!', 'error');
+        return;
+      }
+
+      setCart((prevCart) => {
+        const existingItem = prevCart.find((item) => item.product.id === product.id);
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
+            return prevCart;
+          }
+
+          return prevCart.map((item) =>
+            item.product.id === product.id ? { ...item, quantity: newQuantity } : item
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification('장바구니에 담았습니다', 'success');
+    },
+    [cart, addNotification]
+  );
+
+  const removeFromCart = useCallback((productId: string) => {
+    setCart((prevCart) => prevCart.filter((item) => item.product.id !== productId));
+  }, []);
+
+  const updateQuantity = useCallback(
+    (productId: string, newQuantity: number) => {
+      if (newQuantity <= 0) {
+        removeFromCart(productId);
+        return;
+      }
+
+      const product = products.find((p) => p.id === productId);
+      if (!product) return;
+
+      const maxStock = product.stock;
+      if (newQuantity > maxStock) {
+        addNotification(`재고는 ${maxStock}개까지만 있습니다.`, 'error');
+        return;
+      }
+
+      setCart((prevCart) =>
+        prevCart.map((item) =>
+          item.product.id === productId ? { ...item, quantity: newQuantity } : item
+        )
+      );
+    },
+    [products, removeFromCart, addNotification]
+  );
+
+  const clearCart = useCallback(() => {
+    setCart([]);
+  }, []);
+
+  const getCartTotal = useCallback(
+    (selectedCoupon?: { discountType: 'amount' | 'percentage'; discountValue: number } | null) => {
+      return calculateCartTotal(cart, selectedCoupon);
+    },
+    [cart]
+  );
+
+  return {
+    cart,
+    totalItemCount,
+    addToCart,
+    removeFromCart,
+    updateQuantity,
+    clearCart,
+    getCartTotal,
+  };
+}
