+++ src/basic/hooks/useSearch.ts
@@ -1,0 +1,20 @@
+import { useState, useEffect } from 'react';
+
+export function useSearch(debounceDelay: number = 500) {
+  const [searchTerm, setSearchTerm] = useState('');
+  const [debouncedSearchTerm, setDebouncedSearchTerm] = useState('');
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setDebouncedSearchTerm(searchTerm);
+    }, debounceDelay);
+
+    return () => clearTimeout(timer);
+  }, [searchTerm, debounceDelay]);
+
+  return {
+    searchTerm,
+    setSearchTerm,
+    debouncedSearchTerm,
+  };
+}
